tests/wpt/tests/svg/types/scripted/SVGGraphicsElement.getBBox-06.html

<!DOCTYPE html>
<title>SVGGraphicsElement.prototype.getBBox for non-rendered elements</title>
<link rel="help" href="https://svgwg.org/svg2-draft/coords.html#BoundingBoxes">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 500 500" width="500px" height="500px">
  <symbol id="symbol1">
    <rect id="symbol1_rect" x="50" y="60" width="100" height="150"></rect>
  </symbol>
  <g id="g_none" style="display:none">
    <rect id="g_none_rect" x="70" y="80" width="200" height="250"></rect>
  </g>
  <rect id="none_rect" x="90" y="100" width="20" height="25" display="none"></rect>
</svg>
<script>
  function assert_bbox(id, opt, x, y, width, height, epsilon) {
    if (epsilon == undefined) {
      epsilon = 0.1;
    }
    let bbox = document.getElementById(id).getBBox();
    assert_approx_equals(bbox.x, x, epsilon, id + ".getBBox().x " + JSON.stringify(opt));
    assert_approx_equals(bbox.y, y, epsilon, id + ".getBBox().y " + JSON.stringify(opt));
    assert_approx_equals(bbox.width, width, epsilon, id + ".getBBox().width " + JSON.stringify(opt));
    assert_approx_equals(bbox.height, height, epsilon, id + ".getBBox().height " + JSON.stringify(opt));
  }

  test(() => {
    assert_bbox("symbol1", {}, 0, 0, 0, 0);
    assert_bbox("symbol1_rect", {}, 50, 60, 100, 150);
  }, "Non-rendered symbol with rect");

  test(() => {
    assert_bbox("g_none", {}, 0, 0, 0, 0);
    assert_bbox("g_none_rect", {}, 70, 80, 200, 250);
  }, "display:none <g> with rect");

  test(() => {
    assert_bbox("none_rect", {}, 90, 100, 20, 25);
  }, "display:none <rect>");
</script>
